add <EmptyMessage>
