+++ src/lib/components/EmptyMessage.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+	import { slide } from 'svelte/transition';
+</script>
+
+<p class="align-center flex h-full w-full items-center justify-center text-muted-foreground">
+	<slot></slot>
+</p>
+
